app/requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.38.0
kubernetes==34.1.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
google-cloud-secret-manager==2.25.0